+++ managerClientConsole/alertRclient.py
@@ -13,7 +13,6 @@ from lib import ServerCommunication, ConnectionWatchdog, Receiver
 from lib import SMTPAlert
 from lib import ScreenUpdater
 from lib import Console
-from lib import UpdateChecker
 from lib import GlobalData
 import logging
 import time
@@ -145,31 +144,6 @@ if __name__ == '__main__':
 			smtpToAddr = str(
 				configRoot.find("smtp").find("general").attrib["toAddr"])
 
-		# parse update options
-		updateActivated = (str(
-			configRoot.find("update").find("general").attrib[
-			"activated"]).upper() == "TRUE")
-		if updateActivated is True:
-			updateServer = str(
-				configRoot.find("update").find("server").attrib["host"])
-			updatePort = int(
-				configRoot.find("update").find("server").attrib["port"])
-			updateLocation = str(
-				configRoot.find("update").find("server").attrib["location"])
-			updateCaFile = makePath(str(
-				configRoot.find("update").find("server").attrib["caFile"]))
-			updateInterval = int(
-				configRoot.find("update").find("general").attrib["interval"])
-			updateEmailNotification = (str(
-				configRoot.find("update").find("general").attrib[
-				"emailNotification"]).upper() == "TRUE")
-
-			# email notification works only if smtp is activated
-			if (updateEmailNotification is True
-				and smtpActivated is False):
-				raise ValueError("Update check can not have email "
-					+ "notification activated when smtp is not activated.")
-
 		# get manager settings
 		globalData.description = str(
 			configRoot.find("manager").find("general").attrib[
@@ -267,16 +241,6 @@ if __name__ == '__main__':
 	# => threads terminates when main thread terminates
 	receiver.daemon = True
 	receiver.start()
-
-	# only start update checker if it is activated
-	if updateActivated is True:
-		logging.info("[%s] Starting update check thread." % fileName)
-		updateChecker = UpdateChecker(updateServer, updatePort, updateLocation,
-			updateCaFile, updateInterval, updateEmailNotification, globalData)
-		# set thread to daemon
-		# => threads terminates when main thread terminates
-		updateChecker.daemon = True
-		updateChecker.start()
 
 	logging.info("[%s] Client started." % fileName)