+++ sensorClientLightning/alertRclient.py
@@ -12,7 +12,6 @@ import os
 from lib import ServerCommunication, ConnectionWatchdog
 from lib import SMTPAlert
 from lib import LightningmapSensor, LightningmapDataCollector, SensorExecuter
-from lib import UpdateChecker
 from lib import GlobalData
 import logging
 import time
@@ -137,31 +136,6 @@ if __name__ == '__main__':
 				configRoot.find("smtp").find("general").attrib["fromAddr"])
 			smtpToAddr = str(
 				configRoot.find("smtp").find("general").attrib["toAddr"])
-
-		# parse update options
-		updateActivated = (str(
-			configRoot.find("update").find("general").attrib[
-			"activated"]).upper() == "TRUE")
-		if updateActivated is True:
-			updateServer = str(
-				configRoot.find("update").find("server").attrib["host"])
-			updatePort = int(
-				configRoot.find("update").find("server").attrib["port"])
-			updateLocation = str(
-				configRoot.find("update").find("server").attrib["location"])
-			updateCaFile = makePath(str(
-				configRoot.find("update").find("server").attrib["caFile"]))
-			updateInterval = int(
-				configRoot.find("update").find("general").attrib["interval"])
-			updateEmailNotification = (str(
-				configRoot.find("update").find("general").attrib[
-				"emailNotification"]).upper() == "TRUE")
-
-			# email notification works only if smtp is activated
-			if (updateEmailNotification is True
-				and smtpActivated is False):
-				raise ValueError("Update check can not have email "
-					+ "notification activated when smtp is not activated.")
 
 		# parse all sensors
 		for item in configRoot.find("sensors").iterfind("sensor"):
@@ -290,16 +264,6 @@ if __name__ == '__main__':
 	# => threads terminates when main thread terminates
 	dataCollector.daemon = True
 	dataCollector.start()
-
-	# only start update checker if it is activated
-	if updateActivated is True:
-		logging.info("[%s] Starting update check thread." % fileName)
-		updateChecker = UpdateChecker(updateServer, updatePort, updateLocation,
-			updateCaFile, updateInterval, updateEmailNotification, globalData)
-		# set thread to daemon
-		# => threads terminates when main thread terminates
-		updateChecker.daemon = True
-		updateChecker.start()
 
 	logging.info("[%s] Client started." % fileName)