+++ sensorClientLightning/lib/__init__.py
@@ -11,5 +11,5 @@ from client import ServerCommunication, ConnectionWatchdog, AsynchronousSender
 from smtp import SMTPAlert
 from sensor import LightningmapSensor, \
 	LightningmapDataCollector, SensorExecuter
-from update import UpdateChecker, Updater
+from update import Updater
 from globalData import GlobalData