+++ sensorClientPing/lib/__init__.py
@@ -10,5 +10,5 @@
 from client import ServerCommunication, ConnectionWatchdog, AsynchronousSender
 from smtp import SMTPAlert
 from sensor import PingWatchdogSensor, SensorExecuter
-from update import UpdateChecker, Updater
+from update import Updater
 from globalData import GlobalData